Move the user list client code to its own file
